requirements.txt
streamlit
pandas
gsheetsdb
pandas_datareader
matplotlib
seaborn
newsapi-python